# minor simplification, little bit faster